Avoid contignous disk access in murmur, patch by buggerone
git-svn-id: https://mumble.svn.sourceforge.net/svnroot/mumble/trunk@441 05730e5d-ab1b-0410-a4ac-84af385074fa

+++ ServerDB.cpp
@@ -553,8 +553,10 @@ QList<ServerDB::qpCommand> ServerDB::getCommands() {
 			cmd.second.append(query.value(i));
 		commands << cmd;
 	}
-	query.prepare("DELETE FROM commands");
-	query.exec();
+	if (commands.count() != 0) {
+		query.prepare("DELETE FROM commands");
+		query.exec();
+	}
 	return commands;
 }
